+++ server/src/controllers/specification.ts
@@ -1,25 +1,15 @@
 import { Request, Response } from "express";
 import SpecificationModel from "@/models/specification";
 import { handleError, sendResponse } from "@/lib/utils/helper";
+import { specificationSchema } from "@/schemas/specification";
+import { validateRequest } from "@/config/zod";
 
 // Create Specification
-export const createSpecs = async (req: Request, res: Response) => {
+export const createSpecification = async (req: Request, res: Response) => {
   try {
-    const { items } = req.body;
+    const specification = validateRequest(specificationSchema, req.body);
+    await SpecificationModel.create(specification);
 
-    if (!Array.isArray(items) || items.length === 0)
-      return sendResponse(res, 400, "Items array is required.");
-
-    const addOps = items.map((item: any) => {
-      if (!item.name) {
-        throw new Error("Each item must have a name.");
-      }
-      return SpecificationModel.updateOne({ name: item.name }, item, {
-        upsert: true,
-      });
-    });
-
-    await Promise.all(addOps);
     sendResponse(res, 201, "Specification created successfully.");
   } catch (error) {
     handleError(res, "Error creating specification: ", error);
@@ -27,7 +17,7 @@ export const createSpecs = async (req: Request, res: Response) => {
 };
 
 // Get Specification by ID
-export const getSpecsById = async (req: Request, res: Response) => {
+export const getSpecification = async (req: Request, res: Response) => {
   try {
     const { specsId } = req.params;
     if (!specsId)
@@ -47,7 +37,7 @@ export const getSpecsById = async (req: Request, res: Response) => {
 };
 
 // Get All Specifications
-export const getSpecs = async (_: Request, res: Response) => {
+export const getSpecifications = async (_: Request, res: Response) => {
   try {
     const specifications =
       await SpecificationModel.find().populate("categories");
@@ -60,9 +50,12 @@ export const getSpecs = async (_: Request, res: Response) => {
 };
 
 // Delete Specification
-export const deleteSpecs = async (req: Request, res: Response) => {
+export const deleteSpecification = async (req: Request, res: Response) => {
   try {
     const { specsId } = req.params;
+    if (!specsId)
+      return sendResponse(res, 400, "Specification ID is required.");
+
     const Specs = await SpecificationModel.findByIdAndDelete(specsId);
     if (!Specs) return sendResponse(res, 404, "Specification not found.");
